+++ src/main/java/ladder/domain/PlayerName.java
@@ -1,4 +1,4 @@
-package ladder;
+package ladder.domain;
 
 public class PlayerName {
     private final String name;
